Resolves #10. Specify tick rate

+++ canutils/src/lib.rs
@@ -20,7 +20,7 @@ pub struct Args {
     #[clap(short = 'd', long = "device", value_parser)]
     pub device: String,
     #[clap(short = 't', long = "tick-rate", default_value = "200")]
-    pub tick_rate: u32,
+    pub tick_rate: u64,
 }
 
 pub enum Driver {
@@ -42,5 +42,5 @@ pub enum Command {
 pub struct CommandContext {
     pub socket: CANSocket,
     pub device: String,
-    pub tick_rate: u32,
+    pub tick_rate: u64,
 }
